ingestion_service/requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
confluent-kafka==2.8.0
idna==3.10
kafka-python==2.0.2
requests==2.32.3
simplejson==3.19.3
urllib3==2.3.0
psycopg2-binary==2.9.9